move Logger.h and Logger.cc to the folder Logger
