+++ src/DataStream/Smoothing.cc
@@ -11,8 +11,7 @@
 #include <iostream>
 #include <vector>
 
-#include "../Logger.h"
-
+#include "../Logger/Logger.h"
 #include "Threading.h"
 
 using namespace std;
