+++ src/Moment/MomentGenerator.cc
@@ -6,7 +6,7 @@
 
 #include "MomentGenerator.h"
 
-#include "../Logger.h"
+#include "../Logger/Logger.h"
 
 using namespace carta;
